requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.12
beautifulsoup4==4.12.3
lxml==5.3.0
numpy==1.26.4
faiss-cpu==1.8.0.post1
sentence-transformers==3.0.1
sqlalchemy==2.0.35
httpx==0.27.2
python-dotenv==1.0.1
